New structure: create message and saver packages
