--- src/main/java/com/acme/edu/MessageException.java
@@ -1,11 +1,0 @@
-package com.acme.edu;
-
-/**
- * Created by Java_12 on 30.08.2018.
- */
-public class MessageException extends LogException {
-
-    public MessageException(String message, int code) {
-        super(message, code);
-    }
-}
