+++ src/main/java/com/acme/edu/message/Message.java
@@ -1,4 +1,4 @@
-package com.acme.edu;
+package com.acme.edu.message;
 
 public interface Message {
     public String decorate();
