+++ src/main/java/com/acme/edu/saver/ConsoleSaver.java
@@ -1,6 +1,7 @@
-package com.acme.edu;
+package com.acme.edu.saver;
 
-import java.io.IOException;
+import com.acme.edu.LogException;
+import com.acme.edu.saver.ParameterCheckingSaver;
 
 /**
  * Created by Java_12 on 27.08.2018.
