+++ src/main/java/com/acme/edu/saver/Saver.java
@@ -1,4 +1,6 @@
-package com.acme.edu;
+package com.acme.edu.saver;
+
+import com.acme.edu.LogException;
 
 /**
  * Created by Java_12 on 29.08.2018.
